requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.7
dj-database-url==3.1.2
Django==6.0.4
django-cors-headers==4.9.0
djangorestframework==3.17.1
gunicorn==25.3.0
idna==3.11
packaging==26.1
psycopg2-binary==2.9.11
pycountry==26.2.16
requests==2.33.1
sqlparse==0.5.5
tzdata==2026.1
urllib3==2.6.3
uuid6==2025.0.1
